% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_class_weights}
\alias{check_class_weights}
\title{Check and Set Class Weights}
\usage{
check_class_weights(y, class.weights)
}
\arguments{
\item{y}{The factor of class labels.}

\item{class.weights}{A named numeric vector of class weights or NULL.}
}
\value{
A named numeric vector of class weights.
}
\description{
Internal utility function to validate and set default class weights.
}
\keyword{internal}
