% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make_leaf}
\alias{make_leaf}
\title{Create a Leaf Node}
\usage{
make_leaf(y, class.weights)
}
\arguments{
\item{y}{A factor of class labels.}

\item{class.weights}{Named numeric vector of class weights.}
}
\value{
A list representing a leaf node.
}
\description{
Internal function to create a leaf node with weighted majority vote prediction.
}
\keyword{internal}
