% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.BRFWRF}
\alias{predict.BRFWRF}
\title{Predict Method for Weighted Random Forest / Balanced Random Forest}
\usage{
\method{predict}{BRFWRF}(object, newdata, type = c("class", "prob"))
}
\arguments{
\item{object}{A \code{wrf} or a \code{brf} model object.}

\item{newdata}{A data frame or matrix of new observations.}

\item{type}{Character, "class" for predicted classes or "prob" for class probabilities.}
}
\value{
If \code{type = "class"}, a factor vector of predicted classes.
If \code{type = "prob"}, a matrix of class probabilities.
}
\description{
Generates predictions (classes or probabilities) for new data using a fitted \code{wrf} / \code{brf} model.
}
\examples{
\dontrun{
# Using the wrf model from the previous example:
preds <- predict(model, newdata = iris[, -5], type = "class")
head(preds)
}
}
