% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{predict_tree}
\alias{predict_tree}
\title{Predict with a Single Tree}
\usage{
predict_tree(tree, newdata, type = c("class", "prob"))
}
\arguments{
\item{tree}{A tree object as returned by \code{grow_tree}.}

\item{newdata}{Data frame or matrix of observations.}

\item{type}{"class" or "prob".}
}
\value{
If \code{type = "class"}, a factor vector of predictions. If \code{type = "prob"}, a matrix of class probabilities.
}
\description{
Internal function to predict classes or probabilities from a single tree.
}
\keyword{internal}
