% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{weighted_gini}
\alias{weighted_gini}
\title{Weighted Gini Calculation}
\usage{
weighted_gini(y, class.weights)
}
\arguments{
\item{y}{A factor of class labels.}

\item{class.weights}{Named numeric vector of class weights.}
}
\value{
A numeric value of weighted Gini impurity.
}
\description{
Internal function to compute the weighted Gini impurity.
}
\keyword{internal}
